+++ rollup.config.js
@@ -5,10 +5,9 @@ import json from 'rollup-plugin-json'
 import nodeResolve from 'rollup-plugin-node-resolve'
 import postcss from 'rollup-plugin-postcss'
 import pugPlugin from 'rollup-plugin-pug'
+import { terser } from 'rollup-plugin-terser'
 import typescriptPlugin from 'rollup-plugin-typescript'
-import { uglify } from 'rollup-plugin-uglify'
 import typescript from 'typescript'
-import { minify } from 'uglify-es'
 import { dependencies } from './package.json'
 
 const external = [
@@ -32,7 +31,7 @@ const plugins = [
     plugins: [autoprefixer(), cssnano({ preset: 'default' })],
   }),
   pugPlugin(),
-  !process.env.ROLLUP_WATCH && uglify({}, minify),
+  !process.env.ROLLUP_WATCH && terser(),
 ]
 
 export default [