+++ src/templates/bespoke/progress.ts
@@ -5,7 +5,7 @@ const query = '.bespoke-progress-parent > .bespoke-progress-bar'
 export default function basepokeProgress(deck) {
   deck.on('activate', e => {
     document.querySelectorAll<HTMLElement>(query).forEach(bar => {
-      bar.style.flexBasis = (e.index * 100) / (deck.slides.length - 1) + '%'
+      bar.style.flexBasis = `${(e.index * 100) / (deck.slides.length - 1)}%`
     })
   })
 }