+++ src/file.ts
@@ -3,6 +3,7 @@ import getStdin from 'get-stdin'
 import globby from 'globby'
 import mkdirp from 'mkdirp'
 import path from 'path'
+import { tmpName } from 'tmp'
 
 const markdownExtensions = ['*.md', '*.mdown', '*.markdown', '*.markdn']
 
@@ -37,10 +38,6 @@ export class File {
     return File.initialize(output)
   }
 
-  get directory() {
-    return path.dirname(this.absolutePath)
-  }
-
   async load() {
     this.buffer =
       this.buffer ||
@@ -58,22 +55,52 @@ export class File {
   async save() {
     switch (this.type) {
       case FileType.File:
-        await new Promise<void>((resolve, reject) =>
-          mkdirp(this.directory, e => (e ? reject(e) : resolve()))
-        )
-        await new Promise<void>((resolve, reject) =>
-          fs.writeFile(this.path, this.buffer, e => (e ? reject(e) : resolve()))
-        )
+        await this.saveToFile()
         break
       case FileType.StandardIO:
         process.stdout.write(this.buffer!)
     }
   }
 
+  async saveTmpFile(ext?: string): Promise<File.TmpFileInterface> {
+    const path: string = await new Promise<string>((resolve, reject) => {
+      tmpName({ postfix: ext }, (e, path) => (e ? reject(e) : resolve(path)))
+    })
+
+    await this.saveToFile(path)
+
+    return {
+      path,
+      cleanup: async () => {
+        try {
+          await this.cleanup(path)
+        } catch (e) {}
+      },
+    }
+  }
+
+  private async cleanup(path: string) {
+    return new Promise<void>((resolve, reject) =>
+      fs.unlink(path, e => (e ? reject(e) : resolve()))
+    )
+  }
+
   private convertExtension(extension: string): string {
     return path.join(
       path.dirname(this.path),
       `${path.basename(this.path, path.extname(this.path))}.${extension}`
+    )
+  }
+
+  private async saveToFile(savePath: string = this.path) {
+    await new Promise<void>((resolve, reject) =>
+      mkdirp(
+        path.dirname(path.resolve(savePath)),
+        e => (e ? reject(e) : resolve())
+      )
+    )
+    return new Promise<void>((resolve, reject) =>
+      fs.writeFile(savePath, this.buffer, e => (e ? reject(e) : resolve()))
     )
   }
 
@@ -103,5 +130,12 @@ export class File {
     const instance = new this(filepath)
     tap(instance)
     return instance
+  }
+}
+
+export namespace File {
+  export interface TmpFileInterface {
+    path: string
+    cleanup: () => Promise<void>
   }
 }